Update lots of dependencies

+++ autoplaylist-backend/server/event-store/src/test/kotlin/com/richodemus/autoplaylist/event/ReplayingEventStoreTest.kt
@@ -15,15 +15,15 @@ import com.richodemus.autoplaylist.eventstore.EventType
 import com.richodemus.autoplaylist.eventstore.UserCreated
 import io.micrometer.core.instrument.Counter
 import io.micrometer.core.instrument.MeterRegistry
-import kotlinx.coroutines.experimental.CoroutineDispatcher
-import kotlinx.coroutines.experimental.CoroutineScope
-import kotlinx.coroutines.experimental.Deferred
-import kotlinx.coroutines.experimental.Dispatchers
-import kotlinx.coroutines.experimental.asCoroutineDispatcher
-import kotlinx.coroutines.experimental.async
-import kotlinx.coroutines.experimental.delay
-import kotlinx.coroutines.experimental.launch
-import kotlinx.coroutines.experimental.runBlocking
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.asCoroutineDispatcher
+import kotlinx.coroutines.async
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Before
 import org.junit.Test
@@ -133,7 +133,7 @@ class ReplayingEventStoreTest
                     }
                 }
                 val results = (1..100).map { index ->
-                    delay(index)
+                    delay(index.toLong())
                     target.readEvents(events.size)
                 }.map { it.await() }
                 results.forEach { assertThat(it).isEqualTo(events) }
